Add easing example app, restructure dirs a little
